refactor apollo3 micro speech
