--- tensorflow/lite/micro/examples/micro_speech/apollo3evb/micro_speech.cmd
@@ -1,14 +1,0 @@
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-file ../../../tools/make/gen/apollo3evb_cortex-m4/bin/micro_speech
-target remote localhost:2331
-load ../../../tools/make/gen/apollo3evb_cortex-m4/bin/micro_speech
-monitor reset
-c
-end
-c
